🔨 REFACTOR: move Slug and Password component to common folder
